anspass: unset GIT_ASKPASS
Issue: LIN1023-3432

After the change [1] introduced, the GIT_ASKPASS variable is set when
first run setup with an outdated credentials and then user will prompt
to remove the outdated one and input the new credentials. But as the
GIT_ASKPASS not overrided [1], the user needs to input credentials each
time when connecting the git server as it always returns and make the
askpass_setup function not continue the expected setup_askpass in
03_wrl_askpass.sh as below.
   if [ "${WINDSHARE_SCHEME:0:7}" = "http://" ] || [ "${WINDSHARE_SCHEME:0:8}" = "https://" ]; then
       if [ -n "$GIT_ASKPASS" ]; then
           echo "INFO: Detected GIT_ASKPASS configuration. Disabling built-in askpass functionality."
           return 0
       fi
   fi

So unset GIT_ASKPASS and the corresponding SSH_ASKPASS to let the user
use the setup_askpass to store the credentials in memory once they input
credentials.

[1] http://lxgit.wrs.com/cgit/wrlinux-x/commit/?id=9beaac2f88e4386166ffc20bbad395b28ab55514

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>